feat: bump version to 3.3.4 and refine graph processing query
- Updated project version in pyproject.toml to 3.3.4.
- Modified GraphStore chunk selection to remove unnecessary `graph_processed_at IS NULL` filter and added explanatory comment.
- Adjusted similarity percent calculation in SafeStore to ensure the value is a float.

+++ safe_store/store.py
@@ -600,7 +600,7 @@ class SafeStore:
             ordered_results = []
             for cid, s in zip(top_chunk_ids, top_scores):
                 res = details_map.get(cid, {})
-                res.update({"chunk_id": cid, "similarity_score": float(s), "similarity_percent": round(((s + 1) / 2) * 100, 2)})
+                res.update({"chunk_id": cid, "similarity_score": float(s), "similarity_percent": float(round(((s + 1) / 2) * 100, 2))})
                 ordered_results.append(res)
             return ordered_results
